requirements.txt
# requirements.txt
flask==2.3.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==1.7.6
tldextract==3.4.4
scapy==2.5.0
requests==2.31.0
python-whois==0.8.0
dnspython==2.4.2